bug fixing

+++ src/main/java/com/pollapp/bean/answer/AnswerNumberData.java
@@ -1,14 +1,16 @@
 package com.pollapp.bean.answer;
 
+import java.math.BigDecimal;
+
 public class AnswerNumberData {
 
-    private double percent;
+    private BigDecimal percent;
 
-    public double getPercent() {
+    public BigDecimal getPercent() {
         return percent;
     }
 
-    public void setPercent(double percent) {
+    public void setPercent(BigDecimal percent) {
         this.percent = percent;
     }
 }
